Initial Commit
This version of the use case is the simplified version after meeting with Dr. Moussavi. Many of the smaller activities were consolidated into larger  use cases and the detail that was deemed extra has been discarded. These details should be recaptured when situations are written for each of these cases.